complete restructure the app internal
