+++ src/lib/injector/index.js
