+++ src/lib/options/enable-middleware-shorthand.js
